добавил реализацию метода addProduct() для корзины

+++ src/main/java/com/example/eshop/model/LineOfBasket.java
@@ -22,5 +22,5 @@ public class LineOfBasket {
     @JoinColumn(name = "product_id")
     private Product product;
     @Column(name = "position_cost")
-    float positionCost;
+    double positionCost;
 }
